requirements.txt
argcomplete==1.9.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
boto3==1.9.225
botocore==1.12.225
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
docutils==0.15.2
durationpy==0.5
Flask==1.1.1
future==0.16.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
hjson==3.0.1
httplib2==0.13.1
hyperlink==19.0.0
icalendar==4.0.3
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lxml==4.4.1
MarkupSafe==1.1.1
oauthlib==3.1.0
parsel==1.5.2
placebo==0.9.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.2
PyYAML==5.1.2
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
s3transfer==0.2.1
Scrapy==1.7.3
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.3
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.1
Twisted==19.7.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
w3lib==1.21.0
Werkzeug==0.15.6
wsgi-request-logger==0.4.6
zappa==0.48.2
zope.interface==4.6.0
